+++ lib/plugins/magic_link/magic_link_plugin.dart
@@ -1,0 +1,2 @@
+export 'magic_link_better_auth.dart';
+export 'magic_link_extension.dart';