+++ lib/plugins/multi_session/multi_session_extension.dart
@@ -1,0 +1,10 @@
+import '../../core/api/better_auth_client.dart';
+import '../../core/flutter_better_auth.dart';
+import 'multi_session_better_auth.dart';
+
+extension MultiSessionBetterAuthExtension on BetterAuthClient {
+  MultiSessionBetterAuth get multiSession => MultiSessionBetterAuth(
+    FlutterBetterAuth.dioClient,
+    baseUrl: FlutterBetterAuth.baseUrl,
+  );
+}