+++ lib/plugins/multi_session/multi_session_plugin.dart
@@ -1,0 +1,2 @@
+export 'multi_session_better_auth.dart';
+export 'multi_session_extension.dart';